services/chatbot/pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "Real Madrid Chatbot"
authors = [
    {name = "Caephas",email = "arinzeobidiegwu@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "mangum (>=0.19.0,<0.20.0)",
    "transformers (>=4.50.2,<5.0.0)",
    "aws-xray-sdk (>=2.14.0,<3.0.0)",
    "pip (>=25.0.1,<26.0.0)",
    "google-generativeai (>=0.8.4,<0.9.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
